# apps/control-plane/generated/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Golem control plane",
    "version": "v1"
  },
  "paths": {
    "/api/v1/runtime/events": {
      "post": {
        "operationId": "controlPlaneRuntimeIngest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "schema_version": {
                    "type": "string",
                    "const": "golem.runtime-signal/v1"
                  },
                  "event_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "event_kind": {
                    "type": "string",
                    "enum": [
                      "project.observed",
                      "session.started",
                      "session.resumed",
                      "session.activity",
                      "session.idle",
                      "session.waiting",
                      "session.metadata_patched",
                      "session.ended",
                      "endpoint.claimed",
                      "endpoint.heartbeat",
                      "endpoint.readiness_changed",
                      "endpoint.released",
                      "capabilities.reported"
                    ]
                  },
                  "producer": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "producer_instance_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "harness": {
                    "type": "string",
                    "enum": [
                      "claude",
                      "codex",
                      "opencode",
                      "pi"
                    ]
                  },
                  "producer_sequence": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "correlation_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "causation_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "deduplication_key": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "owner_fence": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "clocks": {
                    "type": "object",
                    "properties": {
                      "source_observed_at": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "source_event_at": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "received_at": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "materialized_at": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      }
                    },
                    "required": [
                      "source_observed_at",
                      "received_at"
                    ],
                    "additionalProperties": false
                  },
                  "provenance": {
                    "type": "object",
                    "properties": {
                      "source": {
                        "type": "string",
                        "enum": [
                          "adapter",
                          "api",
                          "launcher",
                          "legacy_import",
                          "migration"
                        ]
                      },
                      "evidence_id": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256
                      },
                      "confidence": {
                        "type": "string",
                        "enum": [
                          "verified",
                          "observed",
                          "inferred",
                          "legacy"
                        ]
                      }
                    },
                    "required": [
                      "source",
                      "confidence"
                    ],
                    "additionalProperties": false
                  },
                  "clear_fields": {
                    "maxItems": 64,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 160
                    }
                  },
                  "payload": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "schema_version",
                  "event_id",
                  "event_kind",
                  "producer",
                  "producer_instance_id",
                  "harness",
                  "correlation_id",
                  "deduplication_key",
                  "clocks",
                  "provenance",
                  "clear_fields",
                  "payload"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "durably spooled",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.runtime-ingest-receipt/v1"
                    },
                    "event_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 256
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "spooled",
                        "already_pending"
                      ]
                    }
                  },
                  "required": [
                    "schema_version",
                    "event_id",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "invalid runtime signal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "schema_version",
                    "code",
                    "message",
                    "correlation_id"
                  ],
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.api-error/v1"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1024
                    },
                    "correlation_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "schema_version",
                    "code",
                    "message",
                    "correlation_id"
                  ],
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.api-error/v1"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1024
                    },
                    "correlation_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "runtime ingress unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "schema_version",
                    "code",
                    "message",
                    "correlation_id"
                  ],
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.api-error/v1"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1024
                    },
                    "correlation_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meta": {
      "get": {
        "operationId": "controlPlaneMeta",
        "responses": {
          "200": {
            "description": "metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.control-plane-meta/v1"
                    },
                    "instance_id": {
                      "type": "string",
                      "pattern": "^cpi_[0-9a-f-]{36}$"
                    },
                    "service": {
                      "type": "string",
                      "const": "control-plane"
                    },
                    "projections": {
                      "maxItems": 16,
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "schema_version",
                    "instance_id",
                    "service",
                    "projections"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "schema_version",
                    "code",
                    "message",
                    "correlation_id"
                  ],
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.api-error/v1"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1024
                    },
                    "correlation_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projections/{stream}": {
      "get": {
        "operationId": "controlPlaneProjection",
        "parameters": [
          {
            "name": "stream",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "projection",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.control-plane-projection/v1"
                    },
                    "stream": {
                      "type": "string",
                      "enum": [
                        "runtime.live",
                        "runtime.history",
                        "runtime.diagnostics",
                        "projects",
                        "tracker.tree",
                        "tracker.board",
                        "communication.operations"
                      ]
                    },
                    "resource_revision": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "payload": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "schema_version",
                    "stream",
                    "resource_revision",
                    "payload"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "schema_version",
                    "code",
                    "message",
                    "correlation_id"
                  ],
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.api-error/v1"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1024
                    },
                    "correlation_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/browser/session": {
      "post": {
        "operationId": "controlPlaneBrowserSession",
        "responses": {
          "200": {
            "description": "browser session",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.control-plane-browser-session/v1"
                    },
                    "csrf_token": {
                      "type": "string",
                      "minLength": 24,
                      "maxLength": 256
                    }
                  },
                  "required": [
                    "schema_version",
                    "csrf_token"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "schema_version",
                    "code",
                    "message",
                    "correlation_id"
                  ],
                  "properties": {
                    "schema_version": {
                      "type": "string",
                      "const": "golem.api-error/v1"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1024
                    },
                    "correlation_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}